Move Bindling responsibilities back and dedupe
