+++ Cargo.toml
@@ -9,7 +9,7 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
-bindle = { version = "0.5.0", features = ["client"] }
+bindle = { version = "0.5.0", features = ["client"], default-features = false }
 chrono = "0.4"
 clap = { version = "3.0.0-beta.4" }
 colored = "2.0.0"
@@ -19,6 +19,7 @@ glob = "0.3.0"
 hippo = { git = "https://github.com/deislabs/hippo-client-rust", branch = "main" }
 itertools = "0.10.0"
 mime_guess = { version = "2.0" }
+reqwest = { version = "0.11", features = ["stream"] }
 semver = { version = "0.11", features = ["serde"] }
 serde = {version = "1.0", features = ["derive"]}
 sha2 = "0.9"
