+++ src/bindle_pusher.rs
@@ -1,5 +1,7 @@
 use std::path::Path;
 
+use bindle::standalone::StandaloneRead;
+
 use crate::bindle_utils::BindleConnectionInfo;
 
 pub async fn push_all(
@@ -7,5 +9,8 @@ pub async fn push_all(
     bindle_id: &bindle::Id,
     bindle_connection: &BindleConnectionInfo,
 ) -> anyhow::Result<()> {
-    bindle_connection.push_all(path, bindle_id).await
+    let reader = StandaloneRead::new(&path, bindle_id).await?;
+    let client = bindle_connection.client()?;
+    reader.push(&client).await.map_err(|e| anyhow::anyhow!("Error pushing bindle to server: {}", e))?;
+    Ok(())
 }
